Port tests to ExecEngineUtil in InterpreterUtilTest.scala
