+++ casper/src/test/scala/io/casperlabs/casper/helper/HashSetCasperTestNode.scala
@@ -370,13 +370,16 @@ object HashSetCasperTestNode {
   def simpleEEApi[F[_]: Applicative](): ExecutionEngineService[F] =
     new ExecutionEngineService[F] {
       import ipc._
-      private val zero          = Array.fill(32)(0.toByte)
-      private val key           = Key(Key.KeyInstance.Hash(KeyHash(ByteString.copyFrom(zero))))
-      private val transform     = Transform(Transform.TransformInstance.Identity(TransformIdentity()))
-      private val op            = Op(Op.OpInstance.Read(ReadOp()))
-      private val transforEntry = TransformEntry(Some(key), Some(transform))
-      private val opEntry       = OpEntry(Some(key), Some(op))
-      private val ee            = ExecutionEffect(Seq(opEntry), Seq(transforEntry))
+      private val zero = Array.fill(32)(0.toByte)
+
+      private def getExecutionEffect(deploy: Deploy) = {
+        val key           = Key(Key.KeyInstance.Hash(KeyHash(ByteString.copyFromUtf8(deploy.toProtoString))))
+        val transform     = Transform(Transform.TransformInstance.Identity(TransformIdentity()))
+        val op            = Op(Op.OpInstance.Read(ReadOp()))
+        val transforEntry = TransformEntry(Some(key), Some(transform))
+        val opEntry       = OpEntry(Some(key), Some(op))
+        ExecutionEffect(Seq(opEntry), Seq(transforEntry))
+      }
 
       override def emptyStateHash: ByteString = ByteString.copyFrom(zero)
 
@@ -388,7 +391,7 @@ object HashSetCasperTestNode {
         //regardless of their wasm code. It pretends to have run all the deploys,
         //but it doesn't really; it just returns the same result no matter what.
         deploys
-          .map(_ => DeployResult(10, DeployResult.Result.Effects(ee)))
+          .map(d => DeployResult(10, DeployResult.Result.Effects(getExecutionEffect(d))))
           .asRight[Throwable]
           .pure[F]
